+++ client/src/app/router.jsx
@@ -1,5 +1,6 @@
 import { Navigate, createBrowserRouter } from 'react-router-dom'
 import AppShell from '../layouts/AppShell'
+import AutomationsPage from '../pages/AutomationsPage'
 import InboxPage from '../pages/InboxPage'
 import LeadsPage from '../pages/LeadsPage'
 import NotFoundPage from '../pages/NotFoundPage'
@@ -20,6 +21,10 @@ export const router = createBrowserRouter([
       {
         path: 'leads',
         element: <LeadsPage />,
+      },
+      {
+        path: 'automations',
+        element: <AutomationsPage />,
       },
       {
         path: '*',